debug
debug

+++ katsdpdisp/data.py
@@ -1090,7 +1090,7 @@ class SpeadSDReceiver(threading.Thread):
                                                         self.ig['sd_percspectrumflags'].value.astype(np.uint8), \
                                                         self.ig['sd_blmxdata'].value.astype(np.float32).view(np.complex64).swapaxes(0,1)[:,:,0], \
                                                         self.ig['sd_blmxflags'].value.astype(np.uint8).swapaxes(0,1), \
-                                                        self.ig['frequency'].value.astype(np.uint32) if ('frequency' in self.ig) else 0)
+                                                        self.ig['frequency'].value if ('frequency' in self.ig) else 0)
                         elif (hasdata):
                             data = self.ig['sd_data'].value.swapaxes(0,1)
                             for id in range(data.shape[0]):
